feat: refactor package structure and rename classes for better organization
