+++ docs/specs/api/reissue.md
@@ -1,0 +1,50 @@
+### 요청
+
+- 메서드: `POST`
+- 경로: `/api/auth/reissue`
+- 헤더
+
+   ```
+   Content-Type: application/json
+   ```
+
+- 본문
+
+   ```json
+   {
+       "refreshToken":"refreshrefreshrefreshrefreshrefreshrefreshrefreshrefreshrefreshrefresh"
+   }
+   ```
+
+
+- curl 명령 예시
+
+   ```bash
+  curl -i -X POST 'http://localhost:8080/api/auth/refresh' \
+  -H 'Content-Type: application/json' \
+  -d '{
+      "refreshToken":"refreshrefreshrefreshrefreshrefreshrefreshrefreshrefreshrefreshrefresh"
+  }'
+   ```
+
+### 응답
+
+- 상태코드: `200 OK`
+- 본문
+
+   ```json
+   {
+        "accessToken":"accessaccessaccessaccessaccessaccessaccessaccessaccessaccess",
+        "refreshToken":"refreshrefreshrefreshrefreshrefreshrefreshrefreshrefreshrefreshrefresh"
+   }
+
+   ```
+
+### 테스트
+
+- [ ] 올바른 refresh token으로 요청하면 200을 응답한다
+- [ ] 올바른 refresh token으로 요청하면 새로운 access token과 refresh token을 발급해 응답한다
+- [ ] 요청 토큰의 서명이 잘못된 경우 401 Unauthorized가 발생한다
+- [ ] 요청 body가 누락된 경우 400 Bad Request가 발생한다
+- [ ] 유효하지 않은 refresh token으로 요청하 401 Unauthorize가 발생한다
+- [ ] 전달된 토큰은 유효한 JWT 형식이어야 한다
