+++ src/main/java/org/mandarin/booking/adapter/persist/MemberQueryRepository.java
@@ -1,8 +1,8 @@
-package org.mandarin.booking.persist;
+package org.mandarin.booking.adapter.persist;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.mandarin.booking.domain.Member;
+import org.mandarin.booking.domain.model.Member;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
