--- src/main/java/org/mandarin/booking/adapter/webapi/AuthUseCase.java
@@ -1,5 +1,0 @@
-package org.mandarin.booking.adapter.webapi;
-
-public interface AuthUseCase {
-    TokenHolder login(AuthRequest request);
-}
