--- src/main/java/org/mandarin/booking/adapter/webapi/MemberRegisterer.java
@@ -1,5 +1,0 @@
-package org.mandarin.booking.adapter.webapi;
-
-public interface MemberRegisterer {
-    MemberRegisterResponse register(MemberRegisterRequest request);
-}
