+++ src/main/java/org/mandarin/booking/domain/service/TokenProvider.java
@@ -1,4 +1,4 @@
-package org.mandarin.booking.domain;
+package org.mandarin.booking.domain.service;
 
 public interface TokenProvider {
     String generateToken(String userId, String nickName, long expiration);
